Change the the meaing of Status for Windows Services
Fixes ZPS-846

* Status should show if a service is running or stopped
* Events will show if there is a change

Subclass schema.WinServiceInfo also

Fix datapoints to work with 5.x metrics
