+++ ZenPacks/zenoss/Microsoft/Windows/info.py
@@ -1,6 +1,6 @@
 ##############################################################################
 #
-# Copyright (C) Zenoss, Inc. 2016, all rights reserved.
+# Copyright (C) Zenoss, Inc. 2016, 2017, all rights reserved.
 #
 # This content is made available according to terms specified in
 # License.zenoss under the directory where your Zenoss product is installed.
@@ -23,7 +23,7 @@ class InterfaceInfo(schema.InterfaceInfo, IpInterfaceInfo):
     implements(IInterfaceInfo)
 
 
-class WinServiceInfo(WinServiceInfo):
+class WinServiceInfo(schema.WinServiceInfo, WinServiceInfo):
     implements(IWinServiceInfo)
 
     usermonitor = ProxyProperty('usermonitor')
